add protocol filter & rebuild

+++ core/src/common.rs
@@ -91,7 +91,7 @@ impl Description {
 
 #[derive(Default, Clone)]
 pub struct FileInfo {
-    pub link_type: u16,
+    pub link_type: u32,
     pub file_type: FileType,
     pub start_time: u64,
     pub version: String,
